Added some CSS for animtions/Added temp folder for panel mockups
I added some code to calendar.css for animating screens. I also added a
temp folder to hold pages for different pane mockups. Added an event to
onclick for the calendar days with a submit button to get back to the
calendar. This goes to a separate page now and is just for demonstration
of animation. It will have to be done with ajax or javascript.